+++ public_html/index.php
@@ -1,20 +1,24 @@
+<?php
+session_start();
+?>
+
 <!DOCTYPE html>
 <html lang="hr">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="assets/css/base.css">
+    <link rel="stylesheet" href="/assets/css/base.css">
 </head>
 <body>
     <header>
-        <?php require_once("includes/header.php")?>
+        <?php require_once('includes/header.php'); ?>
     </header>
     <main>
-
+        
     </main>
     <footer>
-        <?php require_once("includes/footer.php")?>
+        <?php require_once("includes/footer.php"); ?>
     </footer>
 </body>
 </html>