+++ public_html/lib/db_login.php
@@ -1,0 +1,6 @@
+<?php
+// Data works only for local development
+const SERVERNAME = 'localhost';
+const USERNAME = 'root';
+const PASSWORD = '';
+const DATABASE = 'bookstore';